+++ src/strategies/index.ts
@@ -334,7 +334,7 @@ import * as solvVoucherClaimable from './solv-voucher-claimable';
 import * as h2o from './h2o';
 import * as dopamine from './dopamine';
 import * as lrcL2SubgraphBalanceOf from './lrc-l2-subgraph-balance-of';
-import * as vestingVoucher from './vesting-voucher';
+import * as erc3525VestingVoucher from './erc3525-vesting-voucher';
 
 const strategies = {
   'forta-shares': fortaShares,
@@ -673,7 +673,7 @@ const strategies = {
   h2o,
   dopamine,
   'lrc-l2-subgraph-balance-of': lrcL2SubgraphBalanceOf,
-  'vesting-voucher': vestingVoucher
+  'erc3525-vesting-voucher': erc3525VestingVoucher
 };
 
 Object.keys(strategies).forEach(function (strategyName) {
